# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GleasonAI API",
    "description": "Prostate biopsy AI review service (thesis FedAvg ensemble). **AI decision support only. Final diagnosis requires a pathologist.**",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Upload a slide (multipart) and start AI analysis",
        "operationId": "upload_case_api_v1_cases_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_case_api_v1_cases_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List cases (filters, sort, pagination)",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "-created_at",
                "created_at",
                "-isup_grade",
                "isup_grade",
                "-p_cspca",
                "p_cspca",
                "patient_code"
              ],
              "type": "string",
              "default": "-created_at",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "grade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grade"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "reviewer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reviewer Id"
            }
          },
          {
            "name": "needs_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needs Review"
            }
          },
          {
            "name": "low_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Low Confidence"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "hospital",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hospital"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CaseSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Case with latest prediction and review history",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "Soft-delete a case (admin)",
        "operationId": "delete_case_api_v1_cases__case_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cases/{case_id}/reviews": {
      "post": {
        "tags": [
          "reviews & reports"
        ],
        "summary": "Record a review decision (pathologist/urologist)",
        "operationId": "create_review_api_v1_cases__case_id__reviews_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews & reports"
        ],
        "summary": "Review history (newest first)",
        "operationId": "list_reviews_api_v1_cases__case_id__reviews_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewOut"
                  },
                  "title": "Response List Reviews Api V1 Cases  Case Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/report.pdf": {
      "get": {
        "tags": [
          "reviews & reports"
        ],
        "summary": "PDF case report",
        "operationId": "report_pdf_api_v1_cases__case_id__report_pdf_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/slides/{case_id}.dzi": {
      "get": {
        "tags": [
          "slide viewer"
        ],
        "summary": "DeepZoom descriptor",
        "operationId": "get_dzi_api_v1_slides__case_id__dzi_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/slides/{case_id}_files/{level}/{col}_{row}.jpeg": {
      "get": {
        "tags": [
          "slide viewer"
        ],
        "summary": "DeepZoom tile (JPEG q=85)",
        "operationId": "get_tile_api_v1_slides__case_id__files__level___col___row__jpeg_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Level"
            }
          },
          {
            "name": "col",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Col"
            }
          },
          {
            "name": "row",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_case_api_v1_cases_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": ".tif/.tiff/.svs, max 2 GB"
          },
          "pseudonym_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Pseudonym Code"
          },
          "hospital": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Hospital"
          },
          "slide_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "pseudonym_code"
        ],
        "title": "Body_upload_case_api_v1_cases_post"
      },
      "CaseCreated": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "status",
          "events_url"
        ],
        "title": "CaseCreated"
      },
      "CaseDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "patient_code": {
            "type": "string",
            "title": "Patient Code"
          },
          "hospital": {
            "type": "string",
            "title": "Hospital"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "progress_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Stage"
          },
          "progress_done": {
            "type": "integer",
            "title": "Progress Done"
          },
          "progress_total": {
            "type": "integer",
            "title": "Progress Total"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "uploaded_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By Name"
          },
          "isup_grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isup Grade"
          },
          "p_cspca": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Cspca"
          },
          "low_confidence": {
            "type": "boolean",
            "title": "Low Confidence",
            "default": false
          },
          "review_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_isup": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Isup"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          },
          "slide_sha256": {
            "type": "string",
            "title": "Slide Sha256"
          },
          "slide_bytes": {
            "type": "integer",
            "title": "Slide Bytes"
          },
          "slide_format": {
            "type": "string",
            "title": "Slide Format"
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "predictions_count": {
            "type": "integer",
            "title": "Predictions Count",
            "default": 0
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewOut"
            },
            "type": "array",
            "title": "Reviews",
            "default": []
          },
          "dzi_url": {
            "type": "string",
            "title": "Dzi Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_code",
          "hospital",
          "status",
          "progress_stage",
          "progress_done",
          "progress_total",
          "error",
          "created_at",
          "finished_at",
          "uploaded_by_name",
          "slide_sha256",
          "slide_bytes",
          "slide_format",
          "dzi_url"
        ],
        "title": "CaseDetail"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "queued",
          "processing",
          "done",
          "failed"
        ],
        "title": "CaseStatus"
      },
      "CaseSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "patient_code": {
            "type": "string",
            "title": "Patient Code"
          },
          "hospital": {
            "type": "string",
            "title": "Hospital"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "progress_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Stage"
          },
          "progress_done": {
            "type": "integer",
            "title": "Progress Done"
          },
          "progress_total": {
            "type": "integer",
            "title": "Progress Total"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "uploaded_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By Name"
          },
          "isup_grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isup Grade"
          },
          "p_cspca": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Cspca"
          },
          "low_confidence": {
            "type": "boolean",
            "title": "Low Confidence",
            "default": false
          },
          "review_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_isup": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Isup"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_code",
          "hospital",
          "status",
          "progress_stage",
          "progress_done",
          "progress_total",
          "error",
          "created_at",
          "finished_at",
          "uploaded_by_name"
        ],
        "title": "CaseSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_CaseSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[CaseSummary]"
      },
...
      "ReviewIn": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "final_isup": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Isup",
            "description": "required when amending"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewIn"
      },
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "prediction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "final_isup": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Isup"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          },
          "reviewer_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "prediction_id",
          "decision",
          "final_isup",
          "comment",
          "created_at"
        ],
        "title": "ReviewOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}